teco-c branch: address compiler warnings
and review earlier 'sticky plasters' for compiler errors.

starting with an ncurses example 'from the book' to facilitate
configure issues. In particular a '-lncurses' '-lgpm' '-lncureses'
repot, and suspected to affect 'Cosmic' (or so)

also, aim to compile both with --with-curses (default) and
--without-curses (which may be better if teco is used in similar
ways than sed; in a very 'terse' (terse is not impolite in my book)