clpars.tec: from the documentation branch
Here, this belongs; probably. It sets up the teco environment from
the inside. The (CL) command line (pars) program already had a number
of save/restore set/reset flag instructions. '256,1ez' make teco use
the unix/linux '^J', also known as '
' today, end-of-line marker.

Other things are on the list. This is to be reviewed from the teco-c
branch's view. For example, the loading of initialisation files. The
environment variables that influence behaviours.